requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Pydantic
pydantic[email]==2.10.4
pydantic-settings==2.7.0

# LangGraph & LangChain
langgraph==0.2.60
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14

# Vanna AI
vanna[chromadb,openai,postgres]==0.7.5

# Web Search
tavily-python==0.5.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
aiosqlite==0.20.0
